dashboard/requirements.txt
flask
psycopg2-binary
python-dotenv
flask_cors